+++ week3/src/Theme.tsx
@@ -1,20 +1,6 @@
 import React from "react";
 import { ThemeProvider } from "styled-components";
-
-const theme = {
-  colors: {
-    mainBlue: "#0085FF",
-    darkGrey: "#353535",
-    pointGreen: "#31F167",
-    magenta: "#FB4CFF",
-  },
-  fonts: ["sans-serif", "Roboto"],
-  fontSizes: {
-    small: "1em",
-    medium: "2em",
-    large: "3em",
-  },
-};
+import theme from "./theme";
 
 const Theme = ({ children }: { children: React.ReactNode }) => {
   return <ThemeProvider theme={theme}>{children}</ThemeProvider>;
