+++ week3/src/theme.ts
@@ -1,0 +1,15 @@
+const theme = {
+  colors: {
+    mainBlue: "#0085FF",
+    darkGrey: "#353535",
+    pointGreen: "#31F167",
+    magenta: "#FB4CFF",
+  },
+  fonts: ["sans-serif", "Roboto"],
+  fontSizes: {
+    small: "1em",
+    medium: "2em",
+    large: "3em",
+  },
+};
+export default theme;
